+++ src/components/map-cell.tsx
