+++ src/components/player.tsx
